+++ bpm.h
@@ -37,22 +37,22 @@ class BpmCounter
    unsigned4 measured;
    unsigned4 sample_length;
    stereo_sample2* block;
-   double* audio;
+   float8* audio;
    unsigned4 b1, b2;
-   double * en;
-   double * ts;
-   double * co;
+   float8 * en;
+   float8 * ts;
+   float8 * co;
    fftw_plan forward;
    fftw_plan backward;
    int      input_rate;
    int      measure_rate;
    FILE   * log;
  public:
-   BpmCounter(FILE * l, stereo_sample2 *block, unsigned4 sample_length, int wavrate, double lower_boundary, double higher_boundary) 
+   BpmCounter(FILE * l, stereo_sample2 *block, unsigned4 sample_length, int wavrate, float8 lower_boundary, float8 higher_boundary) 
      {log = l;
      init(sample_length,block,wavrate, lower_boundary,higher_boundary);};
    ~BpmCounter();
-   void init(unsigned4 sample_length, stereo_sample2 *block, int wavrate, double lower_boundary, double higher_boundary);
-   double measure();
+   void init(unsigned4 sample_length, stereo_sample2 *block, int wavrate, float8 lower_boundary, float8 higher_boundary);
+   float8 measure();
 };
 #endif // __loaded__bpm_h__
